add cbuffers, one for tx one for rx, next stateMachine
the cbuffers just works, now it's time to introduce smachine to manaage
in a cooperative way the task.. I've research for preemptuve rtos, but
none of the market brands like freeRtos, nor ti-tos, convince me.. may
be in another aprouch. lets begin with a simple coooperative priorized
smachine evented model.