state machine+events working!
Ready the sm plus events using the cbuffer as a module with variable
data length. Take in account that C2000 minumun data width is 16bits!
yes, there is not such things like a char.. ther is, but the
architecture is based on 16 bits. So sizeof(int)=1 not 2... it's a
detail taht could crash your head some time...
lets go with serial parsere ans some leds effects...