systick with flag UP!, the problem is related to isr chained
I remove the enable/disable isr for timer2 when in smachine and the
problem of jitter goes. It seems that timer2 doesn't like it. I decided
for now add at the end of sm a systick function that pool a flag that
timer2 isr set. So I don't need to disable/enable irqs to read the
events, 'cose til now the event writes are done allways in user mode,
not isr.. but it could change latter..