parser + cmd line MED! receive and response works
Ok, it's really nice feature, it's a menu, a cmd line parser, that
attend [tener] and proccess the line with arguments, like bash style,
but with minumun resources, but I implemented my own printd, so it fit
on ram, and couls print float partially, string, hex, and could adds
more.
 It need some improvement, so I'll finish it and add some features to
report adc to serial for example. let's do it