EQEP MED. So, it works, I'm understandind the calculations
I've manage the pwm1 to simulate the eqep signals and some hardware to
interconect pwm1 outputs to eqep input to simulate the motor encoder.
Noe it works and make some calcs right but I'm trying to undestand the
calculations and adjust the parametersa and word size of parameters to
match my requeriments. So I'm destroying the structures and the eqep
structures right now, let see.