Beginig the porting from launchpad -> controlBoard, it works!
I've port the project simply by undefined the LAUNC macro, and modifying
the make scripts, and it works. So I'am adding some code from demo to my
actual RTOS in the corresponding module, not in a plain file, and I'll
test step by step each of them to understand the whole things.