merging example project with own.. it works!
just spend a day triyng to make the project local to git and not linked
to the ti libs. I need it to make it portable and to understand what is
needed and what not, and not all includede in the example project was
needad clearly. I'm dealing now to organize the tree and be prepared to
destroy the example and combine it's functionality with mine, let see
.*#