add bode plot and testing the prototypes
I'm testing the prototype it works fine but slow, 0.2m/seg and it should
be nice 15m/seg. so I'm 2 magnitude order of the goal...
in other terms, I've manage to test when the motor have stalled and I'm
using it to the sender script so I dont have to wait time, the
controller respond stall when it is stall and the sender script procces
the next line.