PID calibrating and customizing I've made my own PID
I've to change how the TI lib is doing the PID, and now I've change a
little the calculations and also add pid to position if not it cant
follow correctly. and by the way take in account that the speed has not
a lot of preccision mainly in los speed, so the derivative form in speed
pid at los speed doesn't help, but position derivative yes.
ok, now its working somthing good, and I'm commiting to test it with an
CNC controller, cross fingers