added buffers and leds 2 igbt. added xtal for PHY's
I've added one xtal for each PYH instead/appart of the uC pin to drive
the clk for thouse IC's. If the uC could manage to clk them, just
depoppulate the xtal, if not, 3 xtal will be needed. But I guess is
better than distribuiton of 25Mhz between IC's, even they were
buffered...